New MERS estimate API
