+++ src/api/mers/legacy-mers-typedefs.ts
@@ -1,4 +1,4 @@
-export const mersTypedefs = `
+export const legacyMersTypedefs = `
   type MersEstimate {
     id: String!
     seroprevalence: Float!
@@ -17,13 +17,6 @@ export const mersTypedefs = `
     sourceType: String!
     sourceTitle: String!
     insitutution: String
-  }
-
-  enum MersEstimateType {
-    HUMAN_SEROPREVALENCE
-    ANIMAL_SEROPREVALENCE
-    HUMAN_VIRAL
-    ANIMAL_VIRAL
   }
 
   interface MersEstimateInterface {
@@ -281,20 +274,6 @@ export const mersTypedefs = `
 
   union MersEstimate_V2 = HumanMersEstimate | AnimalMersEstimate | HumanMersViralEstimate |  AnimalMersViralEstimate
 
-  enum MersDiagnosisStatus {
-    CONFIRMED
-    DENIED
-  }
-
-  enum MersDiagnosisSource {
-    WORLD_ORGANISATION_FOR_ANIMAL_HEALTH
-    WORLD_HEALTH_ORGANIZATION
-    NATIONAL_AUTHORITIES
-    PUBLICATIONS
-    MEDIA
-    FAO_FIELD_OFFICER
-  }
-
   enum MersAnimalType {
     DOMESTIC
     WILD
@@ -306,93 +285,6 @@ export const mersTypedefs = `
     GOAT
     SHEEP
     CATTLE
-  }
-
-  enum MersEventAnimalType {
-    DOMESTIC
-    WILD
-  }
-
-  enum MersEventAnimalSpecies {
-    CAMEL
-    BAT
-    GOAT
-    SHEEP
-    CATTLE
-  }
-
-  enum MersEventType {
-    HUMAN
-    ANIMAL
-  }
-
-  interface MersEventInterface {
-    id: String!
-    type: MersEventType!
-    diagnosisStatus: MersDiagnosisStatus!
-    diagnosisSource: MersDiagnosisSource!
-    country: CountryIdentifiers!
-    state: String!
-    city: String!
-    latitude: Float!
-    longitude: Float!
-    whoRegion: WHORegion
-    unRegion: UNRegion
-    observationDate: String
-    reportDate: String!
-  }
-
-  type AnimalMersEvent implements MersEventInterface {
-    ####### START INTERFACE FIELDS #######
-    id: String!
-    type: MersEventType!
-    diagnosisStatus: MersDiagnosisStatus!
-    diagnosisSource: MersDiagnosisSource!
-    country: CountryIdentifiers!
-    state: String!
-    city: String!
-    latitude: Float!
-    longitude: Float!
-    whoRegion: WHORegion
-    unRegion: UNRegion
-    observationDate: String
-    reportDate: String!
-    ####### END INTERFACE FIELDS #######
-
-    animalType: MersEventAnimalType!
-    animalSpecies: MersEventAnimalSpecies!
-  }
-
-  type HumanMersEvent implements MersEventInterface {
-    ####### START INTERFACE FIELDS #######
-    id: String!
-    type: MersEventType!
-    diagnosisStatus: MersDiagnosisStatus!
-    diagnosisSource: MersDiagnosisSource!
-    country: CountryIdentifiers!
-    state: String!
-    city: String!
-    latitude: Float!
-    longitude: Float!
-    whoRegion: WHORegion
-    unRegion: UNRegion
-    observationDate: String
-    reportDate: String!
-    ####### END INTERFACE FIELDS #######
-
-    humansAffected: Int!
-    humanDeaths: Int!
-  }
-
-  union MersEvent = AnimalMersEvent | HumanMersEvent
-
-  input PartitionedFaoMersEventsInput {
-    partitionKey: Int!
-  }
-
-  type PartitionedFeoMersEventsOutput {
-    partitionKey: Int!
-    mersEvents: [MersEvent!]!
   }
 
   type MersEstimateFilterOptions {
@@ -412,48 +304,9 @@ export const mersTypedefs = `
     testValidation: [String!]!
   }
 
-  type MersFilterOptions {
-    countryIdentifiers: [CountryIdentifiers!]!
-    whoRegion: [WHORegion!]!
-    unRegion: [UNRegion!]!
-  }
-
-  type YearlyFaoCamelPopulationDataEntry {
-    id: String!
-    countryAlphaThreeCode: String!
-    country: CountryIdentifiers!
-    year: Int!
-    camelCount: Int!
-    camelCountPerCapita: Float
-    whoRegion: WHORegion
-    unRegion: UNRegion
-    note: String!
-  }
-
-  input PartitionedYearlyFaoCamelPopulationDataInput {
-    partitionKey: Int!
-  }
-  
-  type PartitionedYearlyFaoCamelPopulationDataOutput {
-    partitionKey: Int!
-    yearlyFaoCamelPopulationData: [YearlyFaoCamelPopulationDataEntry!]!
-  }
-
-  type FaoMersEventFilterOptions {
-    diagnosisSource: [MersDiagnosisSource!]!
-    animalType: [MersEventAnimalType!]!
-    animalSpecies: [MersEventAnimalSpecies!]!
-  }
-
   type Query {
     mersEstimates: [MersEstimate!]!
     mersEstimates_V2: [MersEstimate_V2!]!
-    mersFilterOptions: MersFilterOptions!
     mersEstimatesFilterOptions: MersEstimateFilterOptions!
-    allFaoMersEventPartitionKeys: [Int!]!
-    partitionedFaoMersEvents(input: PartitionedFaoMersEventsInput!): PartitionedFeoMersEventsOutput!
-    faoMersEventFilterOptions: FaoMersEventFilterOptions!
-    yearlyFaoCamelPopulationDataPartitionKeys: [Int!]!
-    partitionedYearlyFaoCamelPopulationData(input: PartitionedYearlyFaoCamelPopulationDataInput!): PartitionedYearlyFaoCamelPopulationDataOutput!
   }
 `
