+++ src/api/mers/mers-camel-population-data-resolvers.ts
@@ -1,0 +1,18 @@
+import { MongoClient } from "mongodb";
+import { QueryResolvers } from "../graphql-types/__generated__/graphql-types";
+
+interface GenerateMersCamelPopulationDataResolversInput {
+  mongoClient: MongoClient;
+}
+
+interface GenerateMersCamelPopulationDataResolversOutput {
+  mersCamelPopulationDataResolvers: { Query: QueryResolvers }
+}
+
+export const generateMersCamelPopulationDataResolvers = (input: GenerateMersCamelPopulationDataResolversInput): GenerateMersCamelPopulationDataResolversOutput => {
+  return {
+    mersCamelPopulationDataResolvers: {
+      Query: {}
+    }
+  }
+}
