+++ src/api/mers/mers-estimate-typedefs.ts
@@ -1,0 +1,24 @@
+export const mersEstimateTypedefs = `
+  enum MersEstimateType {
+    HUMAN_SEROPREVALENCE
+    ANIMAL_SEROPREVALENCE
+    HUMAN_VIRAL
+    ANIMAL_VIRAL
+  }
+
+  type MersSubEstimate {
+    id: String!
+  }
+
+  type MersPrimaryEstimate {
+    id: String!
+    type: MersEstimateType!
+    geographicalAreaSubestimates: [MersSubEstimate!]!
+    ageSubestimates: [MersSubEstimate!]!
+    sexSubestimates: [MersSubEstimate!]!
+  }
+
+  type Query {
+    mersPrimaryEstimates: [MersPrimaryEstimate!]!
+  }
+`
