+++ src/api/mers/shared-mers-typedefs.ts
@@ -1,0 +1,11 @@
+export const sharedMersTypedefs = `
+  type MersFilterOptions {
+    countryIdentifiers: [CountryIdentifiers!]!
+    whoRegion: [WHORegion!]!
+    unRegion: [UNRegion!]!
+  }
+
+  type Query {
+    mersFilterOptions: MersFilterOptions!
+  }
+`
