+++ src/api/sarscov2/sars-cov-2-typedefs.ts
