+++ src/api/team/team-resolvers.ts
@@ -1,6 +1,6 @@
 import { MongoClient } from "mongodb";
-import { QueryResolvers, TeamMemberSymbol } from "./graphql-types/__generated__/graphql-types.js";
-import { TeamMemberDocument } from "../storage/types";
+import { QueryResolvers, TeamMemberSymbol } from "../graphql-types/__generated__/graphql-types.js";
+import { TeamMemberDocument } from "../../storage/types";
 
 interface GenerateTeamResolversInput {
   mongoClient: MongoClient;
